Restore base profile URL behavior
- Add new profile-specific route
